Enable file tree search for SAPL documents
See: https://github.com/heutelbeck/sapl-policy-engine/issues/39
