+++ sapl-pdp-embedded/src/test/java/io/sapl/prp/filesystem/FileSystemPrpUpdateEventSourceTests.java
@@ -18,10 +18,12 @@
 package io.sapl.prp.filesystem;
 
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.mockStatic;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.when;
 
+import java.io.File;
 import java.util.Collections;
 
 import org.junit.jupiter.api.Test;
@@ -43,6 +45,11 @@ class FileSystemPrpUpdateEventSourceTests {
     void testProcessFileEvent() {
         var source = new FileSystemPrpUpdateEventSource("src/test/resources/it/empty", new DefaultSAPLInterpreter());
 
+        var file1 = mock(File.class);
+        when(file1.getAbsolutePath()).thenReturn("/file1.sapl");
+        var file2 = mock(File.class);
+        when(file2.getAbsolutePath()).thenReturn("/file2.sapl");
+
         try (MockedConstruction<ImmutableFileIndex> mocked = Mockito.mockConstruction(ImmutableFileIndex.class,
                 (mock, context) -> {
                     when(mock.afterFileEvent(any())).thenReturn(mock);
@@ -51,7 +58,7 @@ class FileSystemPrpUpdateEventSourceTests {
 
             try (MockedStatic<FileMonitorUtil> mock = mockStatic(FileMonitorUtil.class)) {
                 mock.when(() -> FileMonitorUtil.monitorDirectory(any(), any()))
-                        .thenReturn(Flux.just(new FileCreatedEvent(null), new FileDeletedEvent(null)));
+                        .thenReturn(Flux.just(new FileCreatedEvent(file1), new FileDeletedEvent(file2)));
 
                 var updates = source.getUpdates();
                 StepVerifier.create(updates).expectNextCount(2L).thenCancel().verify();
